backend/comparision-server/requirements.txt
flask
pymongo
opencv-python
numpy
kafka-python